minor spacing alignment

+++ tests/pytorch/test_nn.py
@@ -1517,11 +1517,11 @@ def test_hgt(idtype, in_size, num_heads):
     sorted_y = m(sorted_g, sorted_x, sorted_ntype, sorted_etype, presorted=False)
     assert sorted_y.shape == (g.num_nodes(), head_size * num_heads)
     # mini-batch
-    train_idx = th.randint(0, 100, (10,), dtype=idtype)
+    train_idx = th.randint(0, 100, (10, ), dtype = idtype)
     sampler = dgl.dataloading.NeighborSampler([-1])
     train_loader = dgl.dataloading.DataLoader(g, train_idx.to(dev), sampler,
-                                              batch_size=batch_size, device=dev,
-                                              shuffle=True)
+                                            batch_size=batch_size, device=dev,
+                                            shuffle=True)
     (input_nodes, output_nodes, block) = next(iter(train_loader))
     block = block[0]
     x = x[input_nodes.to(th.long)]
